refactor: decks optimistic update refactor
