+++ src/services/utils/index.ts
@@ -1,1 +1,2 @@
-export * from './change-deck-favorite-status'
+export * from './perform-optimistic-deck-update'
+export * from './perform-optimistic-favorite-status-update'
